+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * small fixes and improvements
 * on repeat new name respects the language and if date is given, adapts the date
   correspondingly, e.g. "my tapas (2025-7)" to "my tapas (2025-10)"
+* added crystallisation state which is displayes while in crystallization time
 
 #### = 1.2.1 (2025-10-13) =